data/dynamic_pipe_line/requirements.txt
pandas
SQLAlchemy
requests
tqdm
python-dotenv
PyMySQL
